+++ lib/eve_online/esi/models/division.rb
@@ -1,0 +1,22 @@
+module EveOnline
+  module ESI
+    module Models
+      class Division < Base
+        def as_json
+          {
+            hangar: hangar,
+            wallet: wallet
+          }
+        end
+
+        def hangar
+          options['hangar']
+        end
+
+        def wallet
+          options['wallet']
+        end
+      end
+    end
+  end
+end